Remove Nipper::Evidence
