+++ lib/nipper/issue.rb
@@ -4,6 +4,10 @@ module Nipper
       @xml = xml_node
     end
 
+    def evidence_tags
+      [:device_name, :device_osversion, :device_type, :findings_table]
+    end
+
     def supported_tags
       [
         :cvss_base, :cvss_base_vector,
@@ -12,7 +16,7 @@ module Nipper
         :ease, :finding, :impact, :nipperv1_ease, :nipperv1_fix,
         :nipperv1_impact, :nipperv1_rating,
         :recommendation, :title
-      ]
+      ] + evidence_tags
     end
 
     def respond_to?(method, include_private = false)
@@ -31,14 +35,19 @@ module Nipper
 
     def process_field_value(method)
       translations_table = {
+        device_name: 'issuedetails/devices/device/@name',
+        device_osversion: 'issuedetails/devices/device/@osversion',
+        device_type: 'issuedetails/devices/device/@type',
         finding: 'section[@ref="FINDING"]/text',
         impact: 'section[@ref="IMPACT"]/text',
         ease: 'section[@ref="EASE"]/text',
-        recommendation: 'section[@ref="RECOMMENDATION"]/text'
+        recommendation: 'section[@ref="RECOMMENDATION"]/text',
       }
 
       if method == :title
         @xml.attr('title')
+      elsif method == :findings_table
+        process_findings_table
       elsif method.to_s.starts_with?('cvss')
         process_cvss_field(method)
       elsif method.to_s.starts_with?('nipperv1')
@@ -62,6 +71,33 @@ module Nipper
       else
         @xml.xpath("./#{translations_table[base_method]}").attr('score')
       end
+    end
+
+    def process_findings_table
+      table_xml = @xml.at_xpath('section[@title="Finding"]/table')
+
+      return unless table_xml
+
+      headings = table_xml.xpath('headings/heading').map { |heading| heading.text }
+      textile_table = ''
+      headings.each do |heading|
+        textile_table << "|_. #{heading} "
+      end
+
+      textile_table << "|\n"
+
+      tablerows = table_xml.xpath('tablebody/tablerow')
+      tablerows.each do |tablerow|
+        items = tablerow.xpath('tablecell/item').map { |item| item.text }
+
+        items.each do |item|
+          textile_table << "| #{item} "
+        end
+
+        textile_table << "|\n"
+      end
+
+      textile_table
     end
 
     def process_nipperv1_field(method)
